Feels pretty close to a solution using Dijkstra's algorithm

+++ Java/run.sh
@@ -5,7 +5,10 @@ if [ "$#" -eq 1 ] ; then
 	INPUT_FILE=$1
 fi
 
-rm *.class
+if [ -f Solution.class ] ; then
+	rm *.class
+fi
+
 javac Solution.java
 if [ "$?" -eq 0 ] ; then 
 	cat $INPUT_FILE | java Solution
